Added screen resolution and fullscreen/windowed stuff in the menu
Can work on the HUD removal later if the HUD and this menu are implemented I can just "disable enable" it then. But now I got nothing to work with in a seperate scene.

Also the brigness slider is 10X more difficult to do because we are using HDRP. As it is not important to the core gameplay I just left it as is. (With the slider working but not doing anything)