HarshConstraintsOnCloth
This is not a full fix sadly. but with some tweeks I can essentially work around the bug, I have a guess as to what causes it, and it's stupid but it'd require a lot of fenageling to do it right atm. But all in all this should fix it :)